Other refactors
