+++ src/Paraunit/Configuration/Paraunit.php
@@ -17,22 +17,6 @@ class Paraunit
 {
     const PARAUNIT_VERSION = '0.5.1';
 
-    private static $tempDirs = array(
-        '/dev/shm',
-        '/temp',
-    );
-
-    /** @var  string */
-    private $timestamp;
-
-    /**
-     * Paraunit constructor.
-     */
-    public function __construct()
-    {
-        $this->timestamp = uniqid(date('Ymd-His'));
-    }
-
     /**
      * @return ContainerBuilder
      */
@@ -42,6 +26,7 @@ class Paraunit
 
         $loader = new YamlFileLoader($containerBuilder, new FileLocator(__DIR__ . '/../Resources/config/'));
         $loader->load('configuration.yml');
+        $loader->load('file.yml');
         $loader->load('output_container.yml');
         $loader->load('parser.yml');
         $loader->load('printer.yml');
@@ -60,45 +45,5 @@ class Paraunit
         $containerBuilder->compile();
 
         return $containerBuilder;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTempDirForThisExecution()
-    {
-        $dir = self::getTempBaseDir() . DIRECTORY_SEPARATOR . $this->timestamp;
-        self::mkdirIfNotExists($dir);
-        self::mkdirIfNotExists($dir . DIRECTORY_SEPARATOR . 'logs');
-        self::mkdirIfNotExists($dir . DIRECTORY_SEPARATOR . 'coverage');
-
-        return $dir;
-    }
-
-    /**
-     * @return string
-     */
-    public static function getTempBaseDir()
-    {
-        foreach (self::$tempDirs as $directory) {
-            if (file_exists($directory)) {
-                $baseDir = $directory . DIRECTORY_SEPARATOR . 'paraunit';
-                self::mkdirIfNotExists($baseDir);
-
-                return $baseDir;
-            }
-        }
-
-        throw new \RuntimeException('Unable to create a temporary directory');
-    }
-
-    /**
-     * @param string $path
-     */
-    private static function mkdirIfNotExists($path)
-    {
-        if ( ! file_exists($path)) {
-            mkdir($path);
-        }
     }
 }
