+++ src/Paraunit/Resources/config/file.yml
@@ -1,0 +1,11 @@
+services:
+  paraunit.file.temp_directory:
+    class: Paraunit\File\TempDirectory
+
+  paraunit.file.cleaner:
+    class: Paraunit\File\Cleaner
+    arguments:
+      - "@paraunit.file.temp_directory"
+    tags:
+      - { name: kernel.event_listener, event: engine_event.before_start, method: purgeCurrentTempDir }
+      - { name: kernel.event_listener, event: engine_event.end, method: purgeCurrentTempDir }
